Move sections definitions to GroupedTasks
