+++ src/components/TaskListItem.js
@@ -96,7 +96,6 @@ const OrderInfo = ({task, color, width}) => {
 
   return (
     <ItemTouchable
-      onPress={() => console.log("order info pressed")}
       style={{
         alignItems: 'center',
         backgroundColor,
